Refactoring Formatters
