+++ release.sh
@@ -3,4 +3,5 @@
 git tag -l | xargs git tag -d
 git fetch
 ## Release Project
-./mvnw release:prepare release:perform deploy -Prelease -s .mvn/settings.xml
+./mvnw release:prepare -Prelease -s .mvn/settings.xml
+./mvnw release:perform deploy -Prelease -s .mvn/settings.xml
