+++ src/main/java/br/com/ppm/commons/StackTraces.java
@@ -16,6 +16,8 @@
  */
 package br.com.ppm.commons;
 
+import br.com.ppm.commons.formatter.StackTraceFormatters;
+
 /**
  * Utility Class to format Java Stack Traces.
  *
